+++ utils.h
@@ -8,7 +8,7 @@
 #define ARRAY_COUNT(s) (sizeof(s) / sizeof(*s))
 
 int file_exists(const char *fn);
-size_t readvint(unsigned char **stream);
+unsigned long readvint(unsigned char **stream);
 unsigned int read32le(const unsigned char *ptr);
 
 #endif /* UTIL_H */
